refactor: rename contributor components to user components and extract data hook
- Rename ContributorCard to UserCard for consistency
- Rename ContributorLoading to UserCardSkeleton
- Rename ContributorDetailsLoader to UserDetailsSkeleton
- Extract contributors data logic into useContributorsData hook
- Add proper barrel exports with index files
- Move data fetching and filtering logic from IndexPage to custom hook
