+++ src/pages/contributors/IndexPage.tsx
@@ -1,156 +1,13 @@
-import { type JSX, useMemo, useCallback } from 'react';
-import { useQuery } from '@tanstack/react-query';
-import type { FctNodeActiveLast24h } from '@/api/types.gen';
-import { fctNodeActiveLast24hServiceListOptions } from '@/api/@tanstack/react-query.gen';
+import { type JSX, useCallback } from 'react';
 import { Container } from '@/components/Layout/Container';
 import { Header } from '@/components/Layout/Header';
-import { ContributorCard } from './components/ContributorCard';
-import { ContributorLoading } from './components/ContributorLoading';
-import type { ContributorClassification } from './components/ContributorCard';
-
-interface Contributor {
-  username: string;
-  clientName: string;
-  classification: ContributorClassification;
-  nodeCount: number;
-  lastSeen: number;
-  locations: Set<string>;
-  primaryCountry: string | null;
-  primaryCountryCode: string | null;
-  primaryCity: string | null;
-  versions: Set<string>;
-  consensusImplementations: Set<string>;
-  countryCount: Map<string, number>;
-}
-
-// Pure utility functions extracted outside component
-function processNodes(nodes: FctNodeActiveLast24h[]): Contributor[] {
-  const contributorMap = new Map<string, Contributor>();
-
-  nodes.forEach(node => {
-    const clientName = node.meta_client_name || 'Unknown';
-    const username = node.username || clientName;
-    const existing = contributorMap.get(username);
-
-    if (existing) {
-      existing.nodeCount++;
-      existing.lastSeen = Math.max(existing.lastSeen, node.last_seen_date_time || 0);
-
-      if (node.meta_consensus_implementation) {
-        existing.consensusImplementations.add(node.meta_consensus_implementation);
-      }
-
-      if (node.meta_client_version) {
-        existing.versions.add(node.meta_client_version);
-      }
-
-      if (node.meta_client_geo_country) {
-        existing.locations.add(node.meta_client_geo_country);
-        const currentCount = existing.countryCount.get(node.meta_client_geo_country) || 0;
-        existing.countryCount.set(node.meta_client_geo_country, currentCount + 1);
-      }
-    } else {
-      const countryCount = new Map<string, number>();
-      if (node.meta_client_geo_country) {
-        countryCount.set(node.meta_client_geo_country, 1);
-      }
-
-      contributorMap.set(username, {
-        username,
-        clientName,
-        classification: (node.classification || 'unclassified') as ContributorClassification,
-        nodeCount: 1,
-        lastSeen: node.last_seen_date_time || 0,
-        locations: new Set(node.meta_client_geo_country ? [node.meta_client_geo_country] : []),
-        primaryCountry: node.meta_client_geo_country || null,
-        primaryCountryCode: node.meta_client_geo_country_code || null,
-        primaryCity: node.meta_client_geo_city || null,
-        versions: new Set(node.meta_client_version ? [node.meta_client_version] : []),
-        consensusImplementations: new Set(
-          node.meta_consensus_implementation ? [node.meta_consensus_implementation] : []
-        ),
-        countryCount,
-      });
-    }
-  });
-
-  // Calculate primary country (most nodes) for each contributor
-  contributorMap.forEach(contributor => {
-    if (contributor.countryCount.size > 0) {
-      const sortedCountries = Array.from(contributor.countryCount.entries()).sort((a, b) => b[1] - a[1]);
-      contributor.primaryCountry = sortedCountries[0][0];
-    }
-  });
-
-  return Array.from(contributorMap.values()).sort((a, b) => b.nodeCount - a.nodeCount);
-}
-
-function getDisplayVersion(versions: Set<string>): string | undefined {
-  if (versions.size === 0) return undefined;
-  if (versions.size === 1) return Array.from(versions)[0];
-  return 'Multi Versions';
-}
+import { UserCard } from './components/UserCard';
+import { UserCardSkeleton } from './components/UserCardSkeleton';
+import { useContributorsData, getDisplayVersion, type Contributor } from './hooks';
 
 export function IndexPage(): JSX.Element {
-  const {
-    data: pubData,
-    error: pubError,
-    isLoading: pubLoading,
-  } = useQuery({
-    ...fctNodeActiveLast24hServiceListOptions({
-      query: {
-        meta_client_name_starts_with: 'pub-',
-        page_size: 1000,
-      },
-    }),
-  });
-
-  const {
-    data: corpData,
-    error: corpError,
-    isLoading: corpLoading,
-  } = useQuery({
-    ...fctNodeActiveLast24hServiceListOptions({
-      query: {
-        meta_client_name_starts_with: 'corp-',
-        page_size: 1000,
-      },
-    }),
-  });
-
-  const {
-    data: ethData,
-    error: ethError,
-    isLoading: ethLoading,
-  } = useQuery({
-    ...fctNodeActiveLast24hServiceListOptions({
-      query: {
-        meta_client_name_starts_with: 'ethpandaops',
-        page_size: 1000,
-      },
-    }),
-  });
-
-  const isLoading = pubLoading || corpLoading || ethLoading;
-  const error = pubError || corpError || ethError;
-
-  // Memoize processed contributor lists to avoid reprocessing on every render
-  const publicContributors = useMemo(
-    () => processNodes(pubData?.fct_node_active_last_24h ?? []),
-    [pubData?.fct_node_active_last_24h]
-  );
-
-  const corporateContributors = useMemo(
-    () => processNodes(corpData?.fct_node_active_last_24h ?? []),
-    [corpData?.fct_node_active_last_24h]
-  );
-
-  const internalContributors = useMemo(
-    () => processNodes(ethData?.fct_node_active_last_24h ?? []),
-    [ethData?.fct_node_active_last_24h]
-  );
-
-  const totalCount = publicContributors.length + corporateContributors.length + internalContributors.length;
+  const { publicContributors, corporateContributors, internalContributors, totalCount, isLoading, error } =
+    useContributorsData();
 
   // Memoize render function to avoid recreation on every render
   const renderContributorSection = useCallback((title: string, contributors: Contributor[]): JSX.Element | null => {
@@ -161,7 +18,7 @@ export function IndexPage(): JSX.Element {
         <h2 className="mb-6 text-xl/7 font-semibold text-foreground">{title}</h2>
         <ul role="list" className="md: grid grid-cols-1 gap-6 md:grid-cols-1 lg:grid-cols-2 xl:grid-cols-3">
           {contributors.map(contributor => (
-            <ContributorCard
+            <UserCard
               key={contributor.clientName}
               username={contributor.username}
               classification={contributor.classification}
@@ -188,7 +45,7 @@ export function IndexPage(): JSX.Element {
           <h2 className="mb-6 text-xl/7 font-semibold text-foreground">Public Contributors</h2>
           <div className="grid grid-cols-1 gap-6 md:grid-cols-1 lg:grid-cols-2 xl:grid-cols-3">
             {Array.from({ length: 9 }).map((_, index) => (
-              <ContributorLoading key={index} />
+              <UserCardSkeleton key={index} />
             ))}
           </div>
         </div>
