--- src/pages/contributors/components/ContributorCard/index.ts
@@ -1,3 +1,0 @@
-export { ContributorCard } from './ContributorCard';
-export type { ContributorCardProps, ContributorClassification } from './ContributorCard.types';
-export { getCountryFlag, getRelativeTime } from './utils';
