--- src/pages/contributors/components/ContributorDetailsLoader/index.ts
@@ -1,1 +1,0 @@
-export { ContributorDetailsLoader } from './ContributorDetailsLoader';
