--- src/pages/contributors/components/ContributorLoading/index.ts
@@ -1,1 +1,0 @@
-export { ContributorLoading } from './ContributorLoading';
