+++ src/pages/contributors/components/UserCardSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { UserCardSkeleton } from './UserCardSkeleton';
