+++ src/pages/contributors/components/UserDetailsSkeleton/index.ts
@@ -1,0 +1,1 @@
+export { UserDetailsSkeleton } from './UserDetailsSkeleton';
