+++ src/pages/contributors/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useContributorsData, getDisplayVersion } from './useContributorsData';
+export type { Contributor, UseContributorsDataReturn } from './useContributorsData';
